requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.12.0
pytorch-transformers>=1.2.0
datasets>=2.0.0
numpy>=1.19.5
pandas>=1.1.5
opencv-python>=4.5.3
matplotlib>=3.4.3
PyYAML>=5.4.1
anytree>=2.12.1
tqdm>=4.62.3
scipy>=1.5.4
scikit-learn>=0.24.2
Pillow>=8.3.2
yacs>=0.1.8
cityscapesScripts>=2.2.4
clint>=0.5.1
fsspec>=2024.10.0
